Change the reader::read function into a Reader class
The Reader constructor can be used to set the std::istream from which the
input is read and the std::ostream to which the prompt is sent. This
makes it easier to unit test, as both the istream and ostream (and their
content) can be controlled by the testing function.
The class can also be used to set the prompt.

Also remove the " " that was added by the reader::read after the prompt. This " " can now be part of the prompt set in the Reader class.