Build tree-sitter-bash files in a custom_command
Having the compilation in the BUILD_COMMAND of the external
project means that the files are compiled each time make is run, no
matter if the source files haven't been updated.
By moving the compilation into a custom_command, and defining the source
files as dependees, the compilation only happens when the sources
update.